requirements.txt
discord.py
pillow
python-dotenv
google-cloud-vision
